move type declarations to separate file
